requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.10.0
    # via django
async-timeout==5.0.1
    # via redis
certifi==2020.6.20
    # via
    #   pyproj
    #   requests
chardet==3.0.4
    # via requests
database-sanitizer==1.1.0
    # via django-sanitized-dump
django==4.2.26
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-redis
    #   djangorestframework
django-cors-headers==3.5.0
    # via -r requirements.in
django-environ==0.5.0
    # via -r requirements.in
django-filter==2.3.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-sanitized-dump==1.2.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-gis
djangorestframework-gis==0.15
    # via -r requirements.in
idna==2.10
    # via requests
lxml==4.8.0
    # via -r requirements.in
owslib==0.25.0
    # via -r requirements.in
psycopg2==2.9.2
    # via -r requirements.in
pyproj==3.3.0
    # via
    #   -r requirements.in
    #   owslib
python-dateutil==2.8.2
    # via owslib
python-memcached==1.59
    # via -r requirements.in
pytz==2023.3.post1
    # via
    #   -r requirements.in
    #   djangorestframework
    #   owslib
pyyaml==6.0.1
    # via
    #   database-sanitizer
    #   django-sanitized-dump
    #   owslib
raven==6.10.0
    # via -r requirements.in
redis==5.0.7
    # via
    #   -r requirements.in
    #   django-redis
requests==2.32.5
    # via
    #   -r requirements.in
    #   django-helusers
    #   owslib
    #   social-auth-core
six==1.16.0
    # via
    #   database-sanitizer
    #   django-sanitized-dump
    #   python-dateutil
    #   python-memcached
sqlparse==0.3.1
    # via django
typing-extensions==4.15.0
    # via asgiref
tzdata==2023.3
    # via -r requirements.in
urllib3==1.25.10
    # via requests
cachetools==6.2.4
    # via django-helusers
charset-normalizer==3.4.4
    # via requests
defusedxml==0.7.1
    # via social-auth-core
deprecation==2.1.0
    # via django-helusers
django-helusers==0.14.4
    # via -r requirements.in
ecdsa==0.19.1
    # via python-jose
oauthlib==3.3.1
    # via social-auth-core
    # via requests-oauthlib
pyasn1==0.6.1
    # via python-jose
# PyJWT==2.10.1 - Installed separately in Dockerfile
# Required by social-auth-core (>=2.10.1)
python-jose==3.5.0
    # via django-helusers
python3-openid==3.2.0
    # via social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
rsa==4.9.1
    # via python-jose
social-auth-app-django==5.0.0
    # via -r requirements.in
social-auth-core==4.8.3
    # via social-auth-app-django
